Allow users to create additional users for their customer
[Finishes #110348972]

+++ app/services/user_service.rb
@@ -1,6 +1,6 @@
 class UserService
-  def get_all_users
-    User.all
+  def get_all_users(current_user)
+    current_user.admin? ? User.all : User.where(customer: current_user.customer)
   end
 
   def create_user(attributes)
@@ -17,7 +17,9 @@ class UserService
     user.save
   end
 
-  def delete_user(user)
+  def delete_user(current_user, user)
+    return false if current_user == user
+
     user.destroy
   end
 end
